+++ src/classes/Num2Word.mjs
