+++ src/i18n/KO.mjs
@@ -1,4 +1,4 @@
-import Num2Word_Base from '../classes/Num2Word';
+import Num2Word_Base from '../classes/Num2Word.mjs';
 
 export default function() {
   Num2Word_Base.call(this);
