+++ src/i18n/UK.mjs
@@ -1,4 +1,4 @@
-import Num2Word_RU from './RU';
+import Num2Word_RU from './RU.mjs';
 
 export default function() {
   Num2Word_RU.call(this);
