+++ src/n2words.mjs
@@ -1,21 +1,21 @@
-import Num2Word_AR from './i18n/AR';
-import Num2Word_CZ from './i18n/CZ';
-import Num2Word_DE from './i18n/DE';
-import Num2Word_DK from './i18n/DK';
-import Num2Word_EN from './i18n/EN';
-import Num2Word_ES from './i18n/ES';
-import Num2Word_FR from './i18n/FR';
-import Num2Word_HE from './i18n/HE';
-import Num2Word_IT from './i18n/IT';
-import Num2Word_KO from './i18n/KO';
-import Num2Word_LT from './i18n/LT';
-import Num2Word_LV from './i18n/LV';
-import Num2Word_NO from './i18n/NO';
-import Num2Word_PL from './i18n/PL';
-import Num2Word_PT from './i18n/PT';
-import Num2Word_RU from './i18n/RU';
-import Num2Word_TR from './i18n/TR';
-import Num2Word_UK from './i18n/UK';
+import Num2Word_AR from './i18n/AR.mjs';
+import Num2Word_CZ from './i18n/CZ.mjs';
+import Num2Word_DE from './i18n/DE.mjs';
+import Num2Word_DK from './i18n/DK.mjs';
+import Num2Word_EN from './i18n/EN.mjs';
+import Num2Word_ES from './i18n/ES.mjs';
+import Num2Word_FR from './i18n/FR.mjs';
+import Num2Word_HE from './i18n/HE.mjs';
+import Num2Word_IT from './i18n/IT.mjs';
+import Num2Word_KO from './i18n/KO.mjs';
+import Num2Word_LT from './i18n/LT.mjs';
+import Num2Word_LV from './i18n/LV.mjs';
+import Num2Word_NO from './i18n/NO.mjs';
+import Num2Word_PL from './i18n/PL.mjs';
+import Num2Word_PT from './i18n/PT.mjs';
+import Num2Word_RU from './i18n/RU.mjs';
+import Num2Word_TR from './i18n/TR.mjs';
+import Num2Word_UK from './i18n/UK.mjs';
 
 /**
  * Converts numbers to their written form.
