+++ src/n2words_EN.mjs
@@ -1,4 +1,4 @@
-import Num2Word from './i18n/EN';
+import Num2Word from './i18n/EN.mjs';
 
 export default function(n) {
   return new Num2Word().toCardinal(n);
