+++ test/i18n/FR.mjs
