+++ test/i18n/NO.mjs
