+++ test/tests.mjs
@@ -1,24 +1,24 @@
 import test from 'ava';
 
-import n2words from '../src/n2words.js';
-import AR from './i18n/AR';
-import CZ from './i18n/CZ';
-import DE from './i18n/DE';
-import DK from './i18n/DK';
-import EN from './i18n/EN';
-import ES from './i18n/ES';
-import FR from './i18n/FR';
-import HE from './i18n/HE';
-import IT from './i18n/IT';
-import KO from './i18n/KO';
-import LT from './i18n/LT';
-import LV from './i18n/LV';
-import NO from './i18n/NO';
-import PL from './i18n/PL';
-import PT from './i18n/PT';
-import RU from './i18n/RU';
-import TR from './i18n/TR';
-import UK from './i18n/UK';
+import n2words from '../src/n2words.mjs';
+import AR from './i18n/AR.mjs';
+import CZ from './i18n/CZ.mjs';
+import DE from './i18n/DE.mjs';
+import DK from './i18n/DK.mjs';
+import EN from './i18n/EN.mjs';
+import ES from './i18n/ES.mjs';
+import FR from './i18n/FR.mjs';
+import HE from './i18n/HE.mjs';
+import IT from './i18n/IT.mjs';
+import KO from './i18n/KO.mjs';
+import LT from './i18n/LT.mjs';
+import LV from './i18n/LV.mjs';
+import NO from './i18n/NO.mjs';
+import PL from './i18n/PL.mjs';
+import PT from './i18n/PT.mjs';
+import RU from './i18n/RU.mjs';
+import TR from './i18n/TR.mjs';
+import UK from './i18n/UK.mjs';
 
 const i18n = {
     "ar": AR,
