+++ webpack.config.js
@@ -4,10 +4,10 @@ const TerserPlugin = require('terser-webpack-plugin');
 module.exports = {
     mode: 'production',
     entry: {
-        'n2words': './src/n2words.js',
-        'n2words.min': './src/n2words.js',
-        'n2words_EN': './src/n2words_EN.js',
-        'n2words_EN.min': './src/n2words_EN.js'
+        'n2words': './src/n2words.mjs',
+        'n2words.min': './src/n2words.mjs',
+        'n2words_EN': './src/n2words_EN.mjs',
+        'n2words_EN.min': './src/n2words_EN.mjs'
     },
     output: {
         library: 'n2words',
